Add code to generate teapots with groundtruth and white background. Fix various bugs with the OpenDR code regarding int/float variables and the shape of scipy sparse matrices. Add line to instantiate OpenDR that works with Macs.

+++ opendr/camera.py
@@ -54,7 +54,7 @@ class RigidTransform(Ch):
         if wrt is self.v:
             rot = cv2.Rodrigues(self.rt.r)[0]
             
-            IS = np.repeat(np.arange(self.v.r.size), 3)
+            IS = np.repeat(np.arange(self.v.r.size), 3).astype(np.int32)
             JS = np.repeat(np.arange(self.v.r.size).reshape((-1,3)), 3, axis=0)
             data = np.vstack([rot for i in range(np.int(self.v.r.size/3))])
             result = sp.csc_matrix((data.ravel(), (IS.ravel(), JS.ravel())))
@@ -99,7 +99,7 @@ class ProjectPoints(Ch):
         elif wrt is self.v:
             rot = cv2.Rodrigues(self.rt.r)[0]
             data = np.asarray(j[:, 3:6].dot(rot), order='C').ravel()
-            IS = np.repeat(np.arange(self.v.r.size*2/3), 3)
+            IS = np.repeat(np.arange(self.v.r.size*2/3), 3).astype(np.int32)
             JS = np.asarray(np.repeat(np.arange(self.v.r.size).reshape((-1,3)), 2, axis=0), order='C').ravel()
             result = sp.csc_matrix((data, (IS, JS)))
             return result
